feat: add inject function to inject ai config into activator.js
